I have mocked up some code to implement my idea for netlist file synching.  The idea is to allow user to create a quasi-permanent link between an .fpc file and an associated netlist file.  This can be done either from within Project>Options or during a File>Import-netlist.  The effects of having such a link:  (1) whenever a part is inserted/reinserted into the board, or whenever a part is assigned a new footprint, the program will check the (resident) netlist file for references to the part.  It will then assign pins from the part to the appropriate nets wherever possible.    (2) Analogously, if a net is (re)introduced onto the board, the program will assign whatever pins it can to that net automatically, based on the netlist file specs.  This feature I hope will provide the sort of functionality that was formerly provided by so-called phantom pins and parts with null shapes (which I hope to phase out).
Speaking of parts with null shapes, in an effort to discourage them, I now have the program automatically generate a dummy shape (currently
just a little silkscreen circle) for parts whose footprints are not found during a netlist file import.  I find parts with null shapes 
suboptimal for a couple of reasons:  they create great hazards for the programmer, and I think they can be puzzling to the user, who is unlikely
to realize that these parts are actually present, lurking in the partlist but invisible on the board.  I would like ultimately to remove all the
code that is already present and intended to handle these null shapes.

A couple of other bugfixes and minor enhancements:

--- hotkeys for changing the routing angle
--- ability to turn off error beeps (from within Tools>Preferences)